Fix smoothscroll to work with zz/zt/zb
This commit does two things:

---

First, existing zt/zb/zz implementation is buggy when 'smoothscroll' are
'wrap' are set. They can lead to inconsistency bugs with regarding to
repainting and cursor states.

For zz, if the user uses `zz` whenever the top line is smooth-scrolled,
the resulting draw will be wrong. For zt, if the line above the current
line is super long (meaning it is so long that the wrapped line covers
more than the entire screen), it will lead to the cursor being in a bad
state. For zb, the exact conditions are harder to reproduce, but
basically it triggers when the top line is smooth scrolled, and second
line is more than one physical line tall.

Most of the reasons why they are broken is because `reset_skipcol()` or
`redraw_later()` aren't called in the right place, so this change will
make sure to call them appropriately.

---

Second, zb/zz currently will not utilize 'smoothscroll' in a
satisfactory fashion. If 'smoothscroll'/'wrap' are set, we should be
able to accurate put the current line at the bottom/middle of the screen
by smooth-scrolling a long wrapped line at the top of the screen. This
change will make those two commands do that.

---

Also, add tests to regression test these features.