configure: Don't send -U_FORTIFY_SOURCE to preprocessor
The default FORTIFY_SOURCE level was raised in Fedora, so new default
options are passed to CFLAGS/CXXFLAGS by default:

-U_FORTIFY_SOURCE -Wp,-U_FORTIFY_SOURCE -Wp,-D_FORTIFY_SOURCE=N

, where N is 3, the current level. Because Vim wants to be build with
level 1 to prevent crashes in cases which exploit buffer overflow but
are safe in the end, `configure` script removes the related flags, but
the latest change is not considered - so `sed` leaves `-Wp,` in `CFLAGS`,
which breaks compilation.

The patch removes the option completely now.