Fixed bug where fold doesn't expand at end of buffer
The bug is described in Github issue #10698.

The bug was a regression caused by patch 7.4.1896. I assume that this
patch was for speed optimization, but there doesn't seem to be any
reasoning provided, so it would take some work to determine how much of
a difference the optimization made and in what cases.
This patch also caused another bug fixed by patch 8.0.0421. In that
patch, the optimization logic was kept and made more complicated by
adding another condition.

My solution is to simply revert the buggy optimization logic instead of
adding yet another condition to it. In theory this can make the program
slower in certain cases, but I haven't done any speed testing.