feat: add a new fillchar, colorcol, to add a character to the colorcolumn.
This commit add support for a new fillchar called 'colorcol'. When
colorcol is set, the colorcolumn will display that character in the
colorcolumn. The default value is ' ', which results in original
behavior.

Likely colorcol, if set to anything at all, will be set to something
like '│' or '┆' to provide a nice, thin line, but it is certainly not
limited to just those. Good ascii alternatives could be ':', '.' or '~'
(to match NonText).

In the case of multiple colorcolumn's the character is displayed in them
all.

If a typed character is in the colorcolumn, then the typed character
gets precedence over the colorcol char.