Fix heap-based buffer overflow in function utfc_ptr2len at mbyte.c
This issue was previously indicated in
https://huntr.dev/bounties/ae933869-a1ec-402a-bbea-d51764c6618e
(see the POC2 part). This is the readable form of POC:
    norm R0
    split other
    norm R000
    norm <0x16><0xF7>
    ball
    norm zW
The real problem is that ":ball" is executed with the Visual
area starting in column 2, and after the windows are rearraged that
position does not exist in the current buffer. So it will be better
to stop the Visual mode when using ":ball".