Trigger incsearch on cmdline paste
Inserting register literally (e.g. `Ctrl-R +` or `Ctrl-R Ctrl-R "`) in
commandline didn't trigger incsearch or hlsearch.

Steps to reproduce:

1. `vim --clean --cmd 'set incsearch hlsearch | let @"="text" | put'`
2. Type `/` to start searching then type `Ctrl-R Ctrl-R "`.
   Unexpectedly, `text` is not highlighted.

I also found that `cmdline_insert_reg()` returns only `CMDLINE_NOT_CHANGED`
or `GOTO_NORMAL_MODE`. So, the checks of the returned value was
redundant and the `goto cmdline_changed;` line was not executed.

This PR fixes them.
(I'm not sure how to add a test for this...)