Support Python 3 stable ABI to allow mixed version interoperatbility
Vim currently supports embedding Python for use with plugins, and the
"dynamic" linking option allows the user to specify a locally installed
version of Python by setting `pythonthreedll`. However, one caveat is
that the Python 3 libs are not binary compatible across minor versions,
and mixing versions can potentially be dangerous (e.g. let's say Vim was
linked against the Python 3.10 SDK, but the user sets `pythonthreedll`
to a 3.11 lib). Usually, nothing bad happens, but in theory this could
lead to crashes, memory corruption, and other unpredictable behaviors.
It's also difficult for the user to tell something is wrong because Vim
has no way of reporting what Python 3 version Vim was linked with.

For Vim installed via a package manager, this usually isn't an issue
because all the dependencies would already be figured out. For prebuilt
Vim binaries like MacVim (my motivation for working on this), AppImage,
and Win32 installer this could potentially be an issue as usually a
single binary is distributed. This is more tricky when a new Python
version is released, as there's a chicken-and-egg issue with deciding
what Python version to build against and hard to keep in sync when a new
Python version just drops and we have a mix of users of different Python
versions, and a user just blindly upgrading to a new Python could lead to
bad interactions with Vim.

Python 3 does have a solution for this problem: stable ABI / limited API
(see https://docs.python.org/3/c-api/stable.html). The C SDK limits the
API to a set of functions that are promised to be stable across
versions. This pull request adds an ifdef config that allows us to turn
it on when building Vim. Vim binaries built with this option should be
safe to freely link with any Python 3 libraies without having the
constraint of having to use the same minor version.

Note: Python 2 has no such concept and this doesn't change how Python 2
integration works (not that there is going to be a new version of Python
2 that would cause compatibility issues in the future anyway).

---

Technical details:
======

The stable ABI can be accessed when we compile with the Python 3 limited
API (by defining `Py_LIMITED_API`). The Python 3 code (in `if_python3.c`
and `if_py_both.h`) would now handle this and switch to limited API
mode. Without it set, Vim will still use the full API as before so this
is an opt-in change.

The main difference is that `PyType_Object` is now an opaque struct that
we can't directly create "static types" out of, and we have to create
type objects as "heap types" instead. This is because the struct is not
stable and changes from version to version (e.g. 3.8 added a
`tp_vectorcall` field to it). I had to change all the types to be
allocated on the heap instead with just a pointer to them.

Other functions are also simply missing in limited API, or they are
introduced too late (e.g. `PyUnicode_AsUTF8AndSize` in 3.10) to it that
we need some other ways to do the same thing, so I had to abstract a few
things into macros, and sometimes re-implement functions like
`PyObject_NEW`.

One caveat is that in limited API, `OutputType` (used for replacing
`sys.stdout`) no longer inherits from `PyStdPrinter_Type` which I don't
think has any real issue other than minor differences in how they
convert to a string and missing a couple functions like `mode()` and
`fileno()`.

Also fixed an existing bug where `tp_basicsize` was set incorrectly for
`BufferObject`, `TabListObject, `WinListObject`.

Technically, there could be a small performance drop, there is a little
more indirection with accessing type objects, and some APIs like
`PyUnicode_AsUTF8AndSize` are missing, but in practice I didn't see any
difference, and any well-written Python plugin should try to avoid
excessing callbacks to the `vim` module in Python anyway.

I only tested limited API mode down to Python 3.7, which seemes to
compile and work fine. I haven't tried earlier Python versions.

WIP:
======

I marked this as WIP because there are a few things that I want to
finish up, but wanted to gather some feedbacks on this PR first.

- Add way to set `Py_LIMITED_API`. Probably a configure.ac option.
- Add `has()` / `:version` indicator that a Python build has been built
  with stable ABI. I'm not sure if `:version` really need to be changed,
  but I'm imagining `+python3/dyn-stable`.
- Add documentation to explain this, and also how to use `has()` to
  query.
- Add CI coverage to exercise this.
- Test out popular Vim plugins written in Python (see below, I would
  welcome some suggestions) to make sure they still work.
- *Maybe*: Add a `v:python3_version`? This can help the user tell what
  version of Python Vim was built against. Useful esp for non-stable-ABI
  mode.
- *Maybe*: In the old non-stable-ABI mode, throw a warning when loading a
  Python 3 DLL that's a different version from the one Vim expects? I
  may punt on this.

---

I only tested on macOS, down to Python 3.7, with both static and dynamic
linking. If other people want to try this on Windows and Linux that
wouldl be appreciated. I ran a couple Python plugins and they seemed to
work fine (e.g. Ultisnips), but I personally don't use any plugins that
use Python so I may not be running into much edge cases.

Note: If you want to test this, just pull the PR and build. You can comment out the `#define
Py_LIMITED_API 0x03080000` line on the top of if_python3.c to get back
the full API behavior. It should pass all tests and run Python plugins
just like before. One way to know you are running in limited mode is to
run `:py3 print(sys.stdout)` which looks a little different.