Add Python 2 syntax file.
This is a copy of runtime/syntax/python.vim last commit that still
supported Python 2. It also supports Python 3, up to 3.5, because this
version is well tested in previous Vim releases. This ensures better
support for straddling code (Python 2 and 3 compatible).

The use of this older Python 2/3.5 syntax can be enforced by setting
"let python_use_python2_syntax = 1".

The work on runtime/syntax/python.vim after this commit was in support
of Python 3 only. This allows us to move forward with other changes of
that syntax file.