Make converted locale files build in macOS, and fix zh_TW.po
Vim currently uses iconv to convert .po files to different encodings to
accommodate Vim versions that don't have iconv. It uses sed to add a
generated file / do-not-edit marker, but that step fails in macOS
because macOS's version of sed is locale aware and gets confused (see
https://stackoverflow.com/questions/19242275/re-error-illegal-byte-sequence-on-mac-os-x).
To fix this, make sure we prepend `LANG=C` before calling sed which
fixes the issue.

I tested the change in Ubuntu Linux and it still works.

Also fix a couple issues with locale files:

- zh_TW.po is currently generated from zh_TW.UTF-8.po but it's not
  configured as such in the Makefile. In fact, it doesn't contain a
  couple updates from the UTF-8 version. Just add it to the Makefile and
  re-generate it as well.
- Minor: ko.po is generated from ko.UTF-8.po, so it should be under
  `$(MOCONVERTED)` instead of `$(MOFILES)`.