runtime/syntax/gpg.vim: sync with gnupg-2.4.0
A number of options have been added since the last update.  Take the
output from `gpg --dump-option-table` and manually filter the commands
which take args from those which do not.

For reference, this is what was run:

    gpg --dump-option-table | awk -F: '{printf "%-30s %s\n", $1, $3}' |
        sort >/tmp/gpg-options
    grep -Evw '(0|64|128|25[68]|512)$' /tmp/gpg-options |
        awk '{print $1}' >/tmp/gpg-options-arg
    awk '/\<0$/ {print $1}' /tmp/gpg-options >/tmp/gpg-options-noarg
    grep -Ew '(128|25[68])$' /tmp/gpg-options |
        awk '{print $1}' >/tmp/gpg-commands

Unlike the gpg options, the gpg commands were manually split for the
`gpgArg` and `gpgArgError` groups, based on man gpg(1) and the `opts`
array in `g10/gpg.c`.