Fix ruler position in cmdline with 'rulerformat' set
Currently, the ruler is drawn in the cmdline relative to the right side of the
last window when 'rulerformat' is set, and is truncated depending on the size of
the window.

While this behaviour makes sense when there is a statusline, it doesn't when we
have the whole width of the cmdline available for use. This behaviour was also
not consistent with how the ruler works when `rulerformat` is not set.