Need to rewind on reading larger files to keep buffer size constant
We need to be careful when changing the header format. Enryption and
Decryption of file contents depends on the same block size.
Unfortunately, with xchacha20v2, the file header changes and adds
additional 20 bytes and we will use that block size to read the first
block.

However, when starting to read an enrypted file, we do not know yet,
whether it uses the xchacha20 or xchacha20v2 encryption method.

We will initially use the slightly larger block size of the xchacha20v2
version to read the file, but when detecting xchacha20 format and the
file did not fit into the 8K + plus metadata, we need to rewind the
file-descriptor by those additional 20 bytes (add_len bytes of
xchacha20v2), to allow to properly decode it.

Also when decoding the block using xchacha20, we need to adjust the
blocksize and remove the 20 bytes again, since those are only there for
the new xchacha20v2 version, but not the old version.