crypt: Add XChaCha20v2 Encryption method
This adds a new Xchacha20v2 Encryption algorithm. This is basically the
same as the existing xchacha20 algorithm, except that it will store the
pwhashing parameters in the encrypted file header.

The libsodium API strongely suggests to store those parameters alongside
the encrypted to file to be able to decrypt files later on.

This is required, because the libsodium API could change those
parameters and in that case it would no longer be possible to decrypt
files that were created before those constants change.

For now, this is just a copy of the xchacha20 definition. We will make
implementation changes in a later commit