Store additional Data in the header
Amend the cryptmethod_T struct and store the size of additional data
that needs to be stored in the crypt header. This will be used later on
to store the libsodium xchachav20 pwhashing parameters in the enrypted
file header.

For all previous methods, this is 0, except for xchacha20v2 version,
which needs to store:

  - long long opslimit (8 bytes) +
  - size_t memlimit (8 bytes) +
  - int algorithm (4 bytes)

which means we need 20 bytes additional storage.