gui: Implement 'guidarkmode'
Allow gvim to follow the system dark mode and allow the user to override
this choice.

The option is already being honored by the GTK UI when 'guioptions'
cointains 'd', but this new option gives more freedom to implement
different strategies (e.g. the `use_bg` one that follows the 'bg' value)
as needed.

Closes #3922