Ensure DEST_BIN is present before installing tool symlinks
In sandboxed builds, when running with "high" parallelism (-j24), it is
possible that some installation phases run before the target directory
exists.

The failure is silently ignored due to the way the Makefile is made.
This makes it possible that packages end-up with *some* of the tools
provided by vim missing.

```
cd .../bin; ln -s vim ex
.../bin/bash: line 1: cd: .../bin: No such file or directory
```

By using `&&`, the error will not be silently ignored. Furthermore, to
prevent spurious errors on install, a dependency on DEST_BIN is added.