{msg} argument of assert_inrange() behaves differently from others
In assert_inrange(), the {msg} argument replaces the default message,
whereas in other assert functions {msg} is prepended to the default
message.

This PR changes the behavior of assert_inrange() to also prepend {msg}
to the default message by using fill_assert_error().

However, the help for all assert function seems to suggest that {msg}
replaces the default message. Maybe this needs change?

This PR is a superset of #12341.