Correctly check if the linux-modules-extra package exists
commit e741f039cf46ccfbdaaa274618810787129d152b re-enabled the
sound-dummy modules in the CI.

Unfortunately, the commit does not correctly check whether the
linux-modules-extra-... package exist, and so the CI throws an error
```
/home/runner/work/_temp/8c235aad-f1bd-4bdc-adcc-308dd3c8d320.sh: line 2: ]: command not found
```
e.g. here: https://github.com/vim/vim/actions/runs/4961614845/jobs/8878622463#step:8:30

So instead, properly check, whether apt-cache shows is able to show the
package description. We do not need the shell test builtin

  if [ ... ]; then ... fi

because `if` can directly check the return value of the given command,
which makes the test slightly easier to understand.