Require quickfix filename to always be valid
The cases where the file name is not valid are usually misparses of
compiler output - e.g., thinking that "make: *** [Makefile" is a
filename. Vim also already requires the file to exist with %O, %P, and
%Q; this patch simply makes the check apply in all cases.

The behavior of opening a garbage buffer with a nonsensical name is
annoying. Vim presumes the user will look for errors in that file, but
Vim already knows the buffer will be empty.