Ignore some Make output in errorformat
Problem: GNU Make issues messages such as

    make[1]: *** [foo/bar/baz/Makefile:210] Error 1

The default 'errorformat' value matches these messages because of the
%f:%l:%m pattern, and creates quickfix list entries that appear valid
but create a new buffer named 'make[1]: *** [foo/bar/baz/Makefile',
which is obviously incorrect.

Solution: Add entries to 'errorformat' to ignore these messages.

The entries added are:

1. %-Gg%\?make[%*\d]: *** [%f:%l:%m
2. %-Gg%\?make: *** [%f:%l:%m

This covers all of the following cases:

    make[1]: *** [foo/bar/baz/Makefile:200] Error 1
    gmake[1]: *** [foo/bar/baz/Makefile:200] Error 1
    make: *** [Makefile:114: all] Error 2
    gmake: *** [Makefile:114: all] Error 2