termdebug: Parse filenames literally
On Windows with a default path-separator of \ a filename
like test.c may occur, which would be parsed by the termdebug plugin as
\t and therefore replaced by a literal tab.

The same may also happen on Unix, but there filenames with a backslash
in its name are much more uncommon, while on Windows this may happen
just because of the default directory separator.

So stop replacing \n and \t by their typical values when parsing
filenames.

Fixes: #12560 and #12550