Make test for #12528 fail properly without the fix
The test for #12528 doesn't fail without the fix because the original
steps to reproduce run Vim with `-u NONE -i NONE` which puts Vim in
Vi-compatible mode where 'cpoptions' contains the "n" flag, whereas
RunVimInTerminal() runs Vim with `--clean` which doesn't add the "n"
flag to 'cpoptions'.

Adding the "n" flag to 'cpoptions' is enough to make the test work
properly. The `:winsize` command is also unnecessary as the zero-width
window created by `:vsplit` and `:vertical resize 0` is enough.