Specify global/local 'undolevels' to avoid unintentional changes.
When saving and restoring 'undolevels', the constructs `&undolevels` and
`:set undolevels` are problematic.

The construct `&undolevels` reads an unpredictable value; it will be the
local option value (if one has been set), or the global option value
(otherwise), making it unsuitable for saving a value for later
restoration.

Similarly, if a local option value has been set for 'undolevels',
temporarily modifying the option via `:set undolevels` changes the local
value as well as the global value, requiring extra work to restore both
values.

Saving and restoring the option value in one step via the construct
`:let &undolevels = &undolevels` appears to make no changes to the
'undolevels' option, but if a local option has been set to a different
value than the global option, it has the unintended effect of changing
the global 'undolevels' value to the local value.

Update the documentation to explain these issues and recommend explicit
use of global and local option values when saving and restoring.  Update
some unit tests to use `g:undolevels`.