Fix perl compilation errors
Commit 1d7caa58e3c87f75a4becbceabbd1af181ace11e slightly changed how to
integrate with Perl 5.37. Unfortunately, it seems it unexpectedly broke
integration with older Perls.

So parly revert part of the patch:

1) GIMME was changed to GIMME_V for Perl5.37. Let's instead keep the GIMME
   macro and define it to GIMME_V for Perl.37 instead.

2) Change the line
   #if (PERL_REVISION == 5) && (PERL_VERSION >= 18)
back to
  #ifdef DYNAMIC_PERL

I did verify that this comiles using Perl 5.32 and Perl 5.37 in dynamic
mode, but for some reason loading the perl.a library fails, so not
hundert percent sure this is correct.

related #12758