Abort opening cmdwin if autocmds screw things up
Problem: Autocmds triggered from opening the cmdwin (in win_split and do_ecmd)
can cause issues such as E199, as the current checks are insufficient.

Solution: Commands executed from the cmdwin apply to the old curwin/buf, so
they should be kept in a "suspended" state; abort if they've changed.
Also abort if cmdwin/buf was tampered with, and check that curwin is correct.
Try to clean up the cmdwin buffer (only if hidden and non-current to simplify
things; the same approach is used when closing cmdwin normally), and add a beep.

It'd be nice to also check that curwin was *really* created by win_split, as
autocommands can change curwin before it returns (so it can't be assumed to be
that of the split); for now, this means that the cmdwin may not be the botwin in
that case, which is probably OK?