Make "[Command Line]" a special buffer name
Problem: E95 is possible if a buffer called "[Command Line]" already exists when
opening the cmdwin. This can also happen if the cmdwin's buffer could not be
deleted when closing.

Solution: Un-name the cmdwin buffer, and give it a special name instead, similar
to what's done for quickfix buffers and for unnamed prompt and scratch buffers.
As a result, BufFilePre/Post are no longer fired when opening the cmdwin.
Add a "command" key to the dictionary returned by getbufinfo() to differentiate
the cmdwin buffer instead.

NOTE: This is technically a breaking change... maybe this needs a different
solution? (Or maybe this issue can be ignored...)

A GitHub search reveals some plugins expect the old behaviour. However, many of
those plugins also do not seem to account for the string being translated, so
they are subtly broken anyway (not withstanding the fact that you can call any
ol' buffer "[Command Line]" too...)