Remember when options where reset because of the paste/binary option
Mention if depending options for 'binary' or 'paste' have been reset
indirectly. Add a test to verify it works.

Also noticed as small bug, that the global option value for expandtab
was not reset when paste option is set, so fix that while at it.

Regarding the test: for some reason, setting paste does somehow affect
how the internal terminal vim renders the output so we cannot directly
use term_getline(), it will always return an empty vim line. Perhaps Vim
does not correctly redraw when paste option is in affect (probably to
make processing faster). Therefore, we will redir the output and
post-process it a bit (removing path components, removing empty lines)
so that it is easier to scrape the actual buffer line later.

fixes: vim/vim#12837